Extract UserVerificationRequirement enum
